% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark geotrellis  functions}
\usage{
benchmark(ncell = c(100, 1000, 10000), times = 100L)
}
\arguments{
\item{ncell}{\code{integer} number of cells in data to use for bencmarking.}

\item{times}{\code{integer} number of replicate runs.}
}
\value{
\code{gt_Benchmark} object.
}
\description{
This benchmarks functions in this package that leverage geotrellis
\url{http://geotrellis.io} and their counterparts in the raster package.
}
\details{
This function returns a \code{gt_Benchmark} object. This is basically
a list of \code{\link[microbenchmark]{microbenchmark}} objects.
}
\seealso{
\code{\link{print.gt_Benchmark}}, \code{\link{plot.gt_Benchmark}}.
}

