% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_aggregate.R
\name{gt_aggregate}
\alias{gt_aggregate}
\title{Aggregate geotrellis raster data}
\usage{
gt_aggregate(x, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{fact}{\code{integer} Aggregation factor expressed
as number of cells in each direction.}

\item{fun}{\code{character} name of function used to aggregate
values. Available functions are 'mean', 'mode', 'median',
'max', and 'min'. Defaults to 'mean'.}
}
\description{
This function aggregates geotrellis raster layer data
\code{\link{gt_RasterLayer}}.
}
\examples{
\dontrun{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
result <- gt_aggregate(g, fact=2, fun='mean')
}
}

