% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_cellStats.R
\name{gt_cellStats}
\alias{gt_cellStats}
\title{Statistics across cells}
\usage{
gt_cellStats(x, stat = c("sum", "mean", "min", "max", "sd"))
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{stat}{\code{character} name of the statistic to calculate. The following
statistics can be used: 'sum', 'mean', 'min', 'max', and 'sd'.}
}
\value{
\code{numeric} object.
}
\description{
Compute statistics using cell data in a \code{\link{gt_RasterLayer}} object.
}
\details{
This function is similar to the \code{\link[raster]{cellStats}} function.
}

