% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_compareRaster.R
\name{gt_compareRaster}
\alias{gt_compareRaster}
\title{Compare geotrellis raster objects}
\usage{
gt_compareRaster(x, y, extent = TRUE, rowcol = TRUE, crs = TRUE,
  res = TRUE, tolerance = 0.1, stopiffalse = TRUE, showwarning = FALSE)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}}.}

\item{y}{\code{\link{gt_RasterLayer}}.}

\item{extent}{\code{logical} should the objects' extent be compared?
Defaults to \code{TRUE}.}

\item{rowcol}{\code{logical} should the objects' dimensionality be compared?
Defaults to \code{TRUE}.}

\item{res}{\code{logical} should the objects' resolution be compared? 
Defaults to \code{TRUE}.}

\item{tolerance}{\code{numeric} This sets the difference between values
that is permissible for them to be considered equal. Defaults to 0.1.}

\item{stopiffalse}{\code{logical} should an error be thrown if one or
more conditions are not met?}

\item{showwarning}{\code{logical} should a warning be thrown if one or
more conditions are not met?}
}
\value{
\code{logical} are all tested parameters the same?
}
\description{
Evaluate whether two \code{\link{gt_RasterLayer}} objects have the same
extent, number of rows and columns, projection, and resolution
(or a subset of these comparisons).
}
\details{
This function is similar to \code{\link[raster]{compareRaster}}
except that is is less rigorous.
}

