% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_disaggregate.R
\name{gt_disaggregate}
\alias{gt_disaggregate}
\title{Disaggregate geotrellis raster data}
\usage{
gt_disaggregate(x, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{fact}{\code{integer} Disaggregation factor expressed
as number of cells in each direction.}

\item{method}{\code{character} name of method used to disaggregate
values. Available methods are 'ngb' for nearest neighbor assignment,
and 'bilinear' for bilinear interpolation. Defaults to 'ngb'.}
}
\description{
This function disaggregates geotrellis raster layer data
\code{\link{gt_RasterLayer}}.
}
\examples{
\dontrun{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
result <- gt_disaggregate(g, fact=2, method='ngb')
}
}

