% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mask.R
\name{gt_mask}
\alias{gt_mask}
\title{Mask geotrellis raster data}
\usage{
gt_mask(x, y, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{...}{not used.}

\item{mask}{\code{\link{gt_RasterLayer}} object.}

\item{maskvalue}{\code{numeric} The value in \code{mask} that indicates the cells of \code{x}
that should become missing data values. Defaults to \code{NA}.}

\item{updatevalue}{\code{numeric} The value that masked out values in \code{x} should becomes. 
Defaults to \code{NA}.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
Mask data in a \code{\link{gt_RasterLayer}} object using another \code{\link{gt_RasterLayer}}.
}
\details{
This function is similar to \code{\link[raster]{mask}}, except that masked values in \code{x}
are always set to missing data values.
}

