% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_aggregate.R, R/gt_disaggregate.R, R/gt_projectRaster.R
\docType{methods}
\name{gt_aggregate,gt_RasterLayer-method}
\alias{gt_aggregate,gt_RasterLayer-method}
\alias{gt_disaggregate,gt_RasterLayer-method}
\alias{gt_projectRaster}
\alias{gt_projectRaster,gt_RasterLayer,CRS-method}
\alias{gt_projectRaster,gt_RasterLayer,gt_RasterLayer-method}
\title{Project geotrellis raster to a new coordinate system}
\usage{
\S4method{gt_aggregate}{gt_RasterLayer}(x, fact, fun = c("mean", "mode",
  "median", "max", "min"))

\S4method{gt_disaggregate}{gt_RasterLayer}(x, fact, method = c("ngb",
  "bilinear"))

gt_projectRaster(from, to, res, method)

\S4method{gt_projectRaster}{gt_RasterLayer,CRS}(from, to, res,
  method = c("bilinear", "ngb"))

\S4method{gt_projectRaster}{gt_RasterLayer,gt_RasterLayer}(from, to,
  method = c("bilinear", "ngb"))
}
\arguments{
\item{method}{\code{character} method used to compute values for the new \code{\link{gt_RasterLayer}}.
Either 'ngb' (nearest neighbor), which is useful for categorical variables, or 'bilinear' 
(bilinear interpolation; the default value), which is appropriate for continuous variables.}

\item{from}{\code{\link{gt_RasterLayer}} object.}

\item{to}{A \code{\link{gt_RasterLayer}} or \code{\link[sp]{CRS}} object with to reproject 
\code{from} to.}

\item{...}{additional arguments.}

\item{res}{\code{numeric} resolution to reproject the data to. This is required when \code{to}
is a \code{\link[sp]{CRS}} object.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
This function reprojects a geotrellis raster data \code{\link{gt_RasterLayer}}
object to a new coordinate system. This function is similar to \code{\link[raster]{projectRaster}}.
}
\details{
Note that--unlike \code{\link[raster]{projectRaster}}--this function does not have
a \code{to} argument. Therefore users cannot specify a template projected raster to reproject
a project \code{\link{gt_RasterLayer}} and ensure that the returned raster has a specific 
extent and dimensionality.
}
\examples{
\dontrun{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
result <- gt_projectRaster(g, sp::CRS('+init=epsg:3395'), res=50000, method='ngb')
}
}

