% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_raster.R
\name{gt_raster}
\alias{gt_raster}
\alias{gt_raster.RasterLayer}
\alias{gt_raster.character}
\alias{gt_raster.default}
\title{Create geotrellis raster object}
\usage{
gt_raster(x, ...)

\method{gt_raster}{character}(x, ...)

\method{gt_raster}{RasterLayer}(x, ...)

\method{gt_raster}{default}(x, ...)
}
\arguments{
\item{x}{input object.}

\item{...}{additional arguments passed to \code{\link[raster]{raster}}.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
This function creates a geotrellis raster object (\code{\link{gt_RasterLayer}}).
}
\details{
This function essentially loads \code{x} into geotrellis as
as raster data using the following methods:
\itemize{

\item{\code{\link[base]{character}} object}{\code{x} is treated as a file path
for raster data loaded directly into the Scala interpreter.}

\item{\code{\link[raster]{RasterLayer}} object}{\code{x} is saved to disk
and loaded into the Scala interpreter.}

\item{otherwise}{an attempt is made to coerce \code{x} to a \code{\link[raster]{RasterLayer}}
object and then it is saved to disk and loaded into the Scala interpreter.} 

}
}
\examples{
# create gt_raster object using a matrix
x <- gt_raster(matrix(1:4, ncol=2))
# create gt_raster object using a raster
x <- gt_raster(raster(matrix(1:4, ncol=2)))
# create gt_raster object using a file path
x <- gt_raster(system.file("external/test.grd", package="raster"))
}
\seealso{
\code{\link[raster]{raster}}.
}

