% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_RasterLayer-methods.R
\docType{methods}
\name{gt_RasterLayer-methods}
\alias{as.matrix,gt_RasterLayer-method}
\alias{as.raster,gt_RasterLayer-method}
\alias{crs,gt_RasterLayer-method}
\alias{extent,gt_RasterLayer-method}
\alias{gt_RasterLayer-methods}
\alias{ncell,gt_RasterLayer-method}
\alias{ncol,gt_RasterLayer-method}
\alias{nrow,gt_RasterLayer-method}
\alias{res,gt_RasterLayer-method}
\alias{values,gt_RasterLayer-method}
\title{gt_RasterLayer methods}
\usage{
\S4method{crs}{gt_RasterLayer}(x)

\S4method{ncell}{gt_RasterLayer}(x)

\S4method{values}{gt_RasterLayer}(x)

\S4method{res}{gt_RasterLayer}(x)

\S4method{extent}{gt_RasterLayer}(x)

\S4method{ncol}{gt_RasterLayer}(x)

\S4method{nrow}{gt_RasterLayer}(x)

\S4method{as.matrix}{gt_RasterLayer}(x)

\S4method{as.raster}{gt_RasterLayer}(x)
}
\arguments{
\item{x}{\code{gt_RasterLayer} object.}
}
\description{
These are convience wrappers for \code{\link{gt_RasterLayer}} objects.
}
\examples{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
ncell(g) # number of cells
ncol(g) # number of columns
nrow(g) # number of rows
nrow(g) # number of rows
crs(g) # coordinate reference system
res(g) # cell resolution
extent(g) # spatial extent
as.matrix(g) # coerce data to matrix object
as.raster(g) # coerce data to matrix object
}
\seealso{
\code{\link{gt_RasterLayer}}
}

