% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_resample.R
\name{gt_resample}
\alias{gt_resample}
\title{Resample geotrellis raster data}
\usage{
gt_resample(x, y, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{y}{\code{\link{gt_RasterLayer}} object with parameters that \code{x}
should be resampled to.}

\item{...}{not used.}

\item{method}{\code{character} method used to compute values for the new \code{\link{gt_RasterLayer}}.
Either 'ngb' (nearest neighbor), which is useful for categorical variables, or 'bilinear' 
(bilinear interpolation; the default value), which is appropriate for continuous variables.}
}
\value{
\code{\code{\link{gt_RasterLayer}} object}.
}
\description{
This function resamples  a geotrellis raster data \code{\link{gt_RasterLayer}}
object to match the extent, resolution, and dimensionality of another
\code{\link[raster]{projectRaster}}.
}
\details{
This function is similar to \code{\link[raster]{resample}}.
}

