% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_zonal.R
\name{gt_zonal}
\alias{gt_zonal}
\title{Zonal statistics}
\usage{
gt_zonal(x, z, fun = c("mean", "sd", "min", "max", "sum"))
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{z}{\code{\link{gt_RasterLayer}} object with codes representing zones.}

\item{fun}{\code{function} to be applied to summarize the values by zone.}
}
\value{
\code{\link[base]{matrix}} object with a value for each zone (unique values in)
\code{z}.
}
\description{
Calculate statistics from a \code{\link{gt_RasterLayer}} object using
zones specified in a second \code{\link{gt_RasterLayer}}.
}
\details{
This function is similar to \code{\link[raster]{zonal}}. Note that unlike
\code{\link[raster]{zonal}}, the argument to \code{fun} cannot be a code{function}
definition.
}

