% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{plot.gt_Benchmark}
\alias{plot.gt_Benchmark}
\title{Plot}
\usage{
plot.gt_Benchmark(x)
}
\arguments{
\item{x}{input object.}
}
\value{
\code{\link{ggplot2}{ggplot2}} object.
}
\description{
This function plots an object.
}

