% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\docType{methods}
\name{plot}
\alias{plot}
\title{Plot}
\arguments{
\item{x}{input object.}

\item{...}{not used.}
}
\value{
\code{\link{ggplot2}{ggplot2}} object.
}
\description{
This function plots an object.
}

