% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{print.gt_Benchmark}
\alias{print.gt_Benchmark}
\title{Print}
\usage{
print.gt_Benchmark(x)
}
\arguments{
\item{x}{input object.}
}
\value{
None. It is used for the side-effect of printing a message.
}
\description{
This function prints an object.
}

