% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\docType{methods}
\name{print}
\alias{print}
\title{Print}
\arguments{
\item{x}{input object.}

\item{...}{not used.}
}
\value{
None. It is used for the side-effect of printing a message.
}
\description{
This function prints an object.
}

