% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_cellStats.R
\name{gt_cellStats}
\alias{gt_cellStats}
\title{Statistics across cells}
\usage{
gt_cellStats(x, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{...}{not used.}
}
\value{
\code{numeric} object.
}
\description{
Compute statistics using cell data in a \code{\link{gt_RasterLayer}} object.
}
\details{
This function calculates the mean, median, mode, 
standard deviation, minimum, and maximum of the values in a raster. This function
is similar to \code{\link[raster]{cellStats}}.
}
\examples{
g <- gt_raster(raster::raster(matrix(c(1:6), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
gt_cellStats(g)
}

