% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_compareRaster.R
\name{gt_compareRaster}
\alias{gt_compareRaster}
\title{Compare geotrellis raster objects}
\usage{
gt_compareRaster(x, y, extent = TRUE, rowcol = TRUE, crs = TRUE,
  res = TRUE, tolerance = 0.1, stopiffalse = TRUE, showwarning = FALSE)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}}.}

\item{y}{\code{\link{gt_RasterLayer}}.}

\item{extent}{\code{logical} should the objects' extent be compared?
Defaults to \code{TRUE}.}

\item{rowcol}{\code{logical} should the objects' dimensionality be compared?
Defaults to \code{TRUE}.}

\item{res}{\code{logical} should the objects' resolution be compared? 
Defaults to \code{TRUE}.}

\item{tolerance}{\code{numeric} This sets the difference between values
that is permissible for them to be considered equal. Defaults to 0.1.}

\item{stopiffalse}{\code{logical} should an error be thrown if one or
more conditions are not met?}

\item{showwarning}{\code{logical} should a warning be thrown if one or
more conditions are not met?}
}
\value{
\code{logical} are all tested parameters the same?
}
\description{
Evaluate whether two \code{\link{gt_RasterLayer}} objects have the same
extent, number of rows and columns, projection, and resolution
(or a subset of these comparisons).
}
\details{
This function is similar to \code{\link[raster]{compareRaster}}
except that is is less rigorous.
}
\examples{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
g2 <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:3395'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
g3 <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=10, xmx=13, ymn=12, ymx=20))
gt_compareRaster(g, g) # same properties
gt_compareRaster(g, g2, stopiffalse=FALSE) # different crs
gt_compareRaster(g, g2, stopiffalse=FALSE) # different extent
}

