% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_crop.R
\name{gt_crop}
\alias{gt_crop}
\title{Crop geotrellis raster data}
\usage{
gt_crop(x, y)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{y}{\code{\link[raster]{Extent}} or \code{\link{gt_RasterLayer}} object.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
Crop returns a geographic subset of an \code{\link{gt_RasterLayer}} object as specified 
by another \code{\link{gt_RasterLayer}} object.
}
\details{
This function is similar to \code{\link[raster]{crop}}, except that \code{y}
must be a \code{\link{gt_RasterLayer}} object and not just any \code{\link[sp]{Spatial}} object.
}

