% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_projectRaster.R
\name{gt_projectRaster}
\alias{gt_projectRaster}
\title{Reproject geotrellis raster data}
\usage{
gt_projectRaster(from, to, ...)
}
\arguments{
\item{from}{\code{\link{gt_RasterLayer}} object.}

\item{...}{additional arguments.}

\item{crs}{\code{\link[sp]{crs}} new coordinate system to project data to.}

\item{method}{\code{character} method used to compute values for the new \code{\link{gt_RasterLayer}}.
Either 'ngb' (nearest neighbor), which is useful for categorical variables, or 'bilinear' 
(bilinear interpolation; the default value), which is appropriate for continuous variables.}
}
\value{
\code{\code{\link{gt_RasterLayer}} object}.
}
\description{
This function reprojects a geotrellis raster data \code{\link{gt_RasterLayer}}
object to a new coordinate system. This function is similar to \code{\link[raster]{projectRaster}}.
}
\details{
Note that--unlike \code{\link[raster]{projectRaster}}--this function does not have
a \code{to} argument. Therefore users cannot specify a template projected raster to reproject
a project \code{\link{gt_RasterLayer}} and ensure that the returned raster has a specific 
extent and dimensionality.
}

