% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_raster.R
\name{gt_raster}
\alias{gt_raster}
\title{Create geotrellis raster object}
\usage{
gt_raster(x, ...)
}
\arguments{
\item{x}{input object.}

\item{...}{additional arguments passed to \code{\link[raster]{raster}} if \code{x}
is not \code{character} or \code{raster}.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
This function creates a geotrellis raster object (\code{\link{gt_RasterLayer}}).
}
\details{
This function essentially loads \code{x} into geotrellis as
as raster data using the following methods:
\itemize{

\item{\code{\link[base]{character}} object}{\code{x} is treated as a file path
for raster data loaded directly into the Scala interpreter.}

\item{\code{\link[raster]{RasterLayer-class}} object}{\code{x} is saved to disk
and loaded into the Scala interpreter.}

\item{otherwise}{an attempt is made to coerce \code{x} to a \code{\link[raster]{RasterLayer-class}}
object and then it is saved to disk and loaded into the Scala interpreter.} 

}
}
\examples{
\dontrun{
# create gt_raster object using a raster
g1 <- gt_raster(raster(matrix(1:4, ncol=2)))
# create gt_raster object using a geotiff path
g2 <- gt_raster(system.file('extdata', 'test.tif', package='geotrellis'))
# create gt_raster object using default method
g3 <- gt_raster(matrix(1:4, ncol=2))
}
}
\seealso{
\code{\link[raster]{raster}}.
}

