% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_RasterLayer.R
\name{gt_RasterLayer}
\alias{gt_RasterLayer}
\title{Geotrellis raster layer class}
\description{
This class represents geotrellis raster data objects in R. Note that
only expert users should attempt to manipulate data using these
methods directly.
}
\section{Usage}{


\preformatted{r <- gt_RasterLayer$new()}

r$print()

r$read_data(path)
r$write_data(path)
r$delete_data()
r$copy_data()

r$values()
r$extent()
r$crs()
}

\section{Arguments}{


\describe{
\item{path}{\code{character} file path for GeoTIFF raster.}
}
}

\section{Details}{


\code{$new()} create a new raster layer object.
\code{$finalize()} destroy the object.
\code{$print()} print the object.
\code{$read_data()} read data to asosciate with the object.
\code{$write_data()} write data to asosciate with the object.
\code{$delete_data()} delete the data associated with the object.
\code{$values()} values in data associated with the object.
\code{$extent()} extent of the data associated with the object.
\code{$crs()} coordinate reference system associated with the object.
}
\seealso{
\code{\link{gt_RasterLayer-methods}}.
}

