% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_resample.R
\name{gt_resample}
\alias{gt_resample}
\title{Resample geotrellis raster data}
\usage{
gt_resample(x, y, method)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{y}{\code{\link{gt_RasterLayer}} object with parameters that \code{x}
should be resampled to.}

\item{...}{not used.}

\item{method}{\code{character} method used to compute values for the new \code{\link{gt_RasterLayer}}.
Either 'ngb' (nearest neighbor), which is useful for categorical variables, or 'bilinear' 
(bilinear interpolation; the default value), which is appropriate for continuous variables.}
}
\value{
\code{\link{gt_RasterLayer}} object.
}
\description{
This function resamples  a geotrellis raster data \code{\link{gt_RasterLayer}}
object to match the extent, resolution, and dimensionality of another
\code{\link[raster]{projectRaster}}.
}
\details{
This function is similar to \code{\link[raster]{resample}}.
}
\examples{
\dontrun{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
template <- gt_raster(raster::raster(matrix(runif(4), ncol=2),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
result <- gt_resample(g, template, method='bilinear')
}
}

