% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_writeRaster.R
\name{gt_writeRaster}
\alias{gt_writeRaster}
\title{Write geotrellis raster to file}
\usage{
gt_writeRaster(x, path)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{x}{\code{character} file path.}
}
\value{
\code{invisible()}. This function is used for its side-effect
of producing files on disk.
}
\description{
Save a \code{\link{gt_RasterLayer}} object to disk in a GeoTIFF format.
}
\examples{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
gt_writeRaster(g, tempfile(fileext='.tif'))
}

