% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_zonal.R
\name{gt_zonal}
\alias{gt_zonal}
\title{Zonal statistics}
\usage{
gt_zonal(x, z, ...)
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{z}{\code{\link{gt_RasterLayer}} object with codes representing zones.}

\item{...}{not used.}
}
\value{
\code{\link[base]{matrix}} object with a value for each zone (unique values in)
\code{z}.
#' @details This function calculates the mean, median, mode, and
standard deviation in each zone. It is similar to \code{\link[raster]{zonal}}.
}
\description{
Calculate statistics from a \code{\link{gt_RasterLayer}} object using
zones specified in a second \code{\link{gt_RasterLayer}}.
}
\examples{
g <- gt_raster(raster::raster(matrix(runif(9), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
z <- gt_raster(raster::raster(matrix(sample(1:3, 9, replace=TRUE), ncol=3),
                              crs=sp::CRS('+init=epsg:4326'),
                              xmn=0, xmx=3, ymn=2, ymx=10))
result <- gt_zonal(g, z)
}

