% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_crop.R, R/gt_zonal.R
\name{gt_zonal}
\alias{gt_zonal}
\title{Crop geotrellis raster data}
\usage{
gt_zonal(x, z, fun = c("mean", "sd", "min", "max", "sum"))

gt_zonal(x, z, fun = c("mean", "sd", "min", "max", "sum"))
}
\arguments{
\item{x}{\code{\link{gt_RasterLayer}} object.}

\item{z}{\code{\link{gt_RasterLayer}} object with codes representing zones.}

\item{fun}{\code{function} to be applied to summarize the values by zone.}

\item{y}{\code{\link{gt_RasterLayer}} object.}

\item{x}{\code{\link{gt_RasterLayer}} object.}
}
\value{
\code{\link{gt_RasterLayer}} object.

\code{\link[base]{matrix}} object with a value for each zone (unique values in)
\code{z}.
}
\description{
Crop returns a geographic subset of an \code{\link{gt_RasterLayer}} object as specified 
by another \code{\link{gt_RasterLayer}} object.

Calculate statistics from a \code{\link{gt_RasterLayer}} object using
zones specified in a second \code{\link{gt_RasterLayer}}.
}
\details{
This function is similar to \code{\link[raster]{crop}}, except that \code{y}
must be a \code{\link{gt_RasterLayer}} object and not just any \code{\link[sp]{Spatial}} object.

This function is similar to \code{\link[raster]{zonal}}. Note that unlike
\code{\link[raster]{zonal}}, the argument to \code{fun} cannot be a code{function}
definition.
}

